requirements.txt
fastapi
uvicorn
openai
python-dotenv
requests
beautifulsoup4
moviepy
Pillow